+++ src/components/Presenter/Main/Feed/FeedContainer.tsx
@@ -6,11 +6,11 @@ import React, { useState } from 'react';
 
 import { TabType } from './type';
 
-import * as ms from '../MainStyled';
-
 import FeedTitle from './FeedTitle';
 import FeedTab from './FeedTab';
 import FeedContent from './FeedContent';
+import styled from '@emotion/styled';
+import { theme } from '@constants/theme';
 
 export default function Feed() {
   const [currentTab, setCurrentTab] = useState<TabType>('lecture');
@@ -23,10 +23,22 @@ export default function Feed() {
   };
 
   return (
-    <ms.FeedContainer>
+    <FeedContainer>
       <FeedTitle title={'나의 피드'} />
       <FeedTab feedList={['오늘의 강의', '학사 일정']} handleTab={handleTab} />
       <FeedContent currentTab={currentTab} />
-    </ms.FeedContainer>
+    </FeedContainer>
   );
 }
+
+const FeedContainer = styled.div`
+  width: 100%;
+  min-height: 220px;
+  height: 30vh;
+  max-height: 280px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  background-color: ${theme.color.main};
+`;
