+++ src/components/Presenter/Main/MainStyled.ts
@@ -12,18 +12,6 @@ export const ContentContainer = styled.div`
   overflow: hidden;
 `;
 
-export const FeedContainer = styled.div`
-  width: 100%;
-  min-height: 220px;
-  height: 30vh;
-  max-height: 280px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  background-color: ${theme.color.main};
-`;
-
 export const Content = styled.div`
   width: 100%;
   flex: 5;
